Fixed local endpoint acquisition for `IHubService` implementation
